commands: remove manually created mocks and adjust tests for newly created mocks via mockery
